Support loading maps for Money.Ecto.Amount.Type
We were using the `Money.Ecto.Amount.Type` type in our embedded
schemas before it was technically supported. It seems that this
functioned by _not_ loading or casting those Money values, meaning
that they were stored as a JSON object instead of an integer as the
`Money.Ecto.Amount.Type` should do.

However, now Money has officially added support for this type in
embedded schemas. This means that when loading those records, Money
assumes that the values have been stored as integers. For us, this is
not the case.

By adding support for loading maps, we create a stop-gap solution
where we can load our historical map records as well as integers, and
all persistence will use the new (correct) integer format. This will
buy us time to run a migration to convert all map records to integers.